requirements.txt
openai
streamlit
python-dotenv
fastapi 
uvicorn 
pydantic
requests